auto:fix(sql-types) align feature hudi_read schema with BIGINT prep
HudiUtilities.validateTypeCompatible rejected `hb_nulls` because the feature
test declared `id int` while the prep script (prepare_hudi/hudi_smoke.sql)
writes `CAST(1 AS BIGINT)`. dlagent raised UnsupportedTypeException
"expected GPDB type BIGINT, actual GPDB type INTEGER" and the query failed.

- hb_nulls: id int -> id bigint (matches prep)
- regen 01_hudi_basic.out: captures fixed schema, MOR merged data
  (Mouse Pro 39.99/2010 + Monitor 4K 499.99/2011), and 5-row COUNT
- regen 02_hudi_features.out: hf_part now returns populated partition rows

+++ contrib/datalake_fdw/automation/sqlrepo/feature/hudi_read/sql/01_hudi_basic.sql
@@ -58,7 +58,7 @@ SELECT id, name FROM hb_mor ORDER BY id;
 -- ============================================================
 SELECT test_log('Test 5: NULL table read');
 
-CREATE FOREIGN TABLE hb_nulls (id int, name text, val int, ts bigint)
+CREATE FOREIGN TABLE hb_nulls (id bigint, name text, val int, ts bigint)
 SERVER hudi_ft_server
 OPTIONS (filePath 'default.hudi_null_table', catalog_type 'hive', server_name 'hive_cluster',
          hdfs_cluster_name 'paa_cluster', table_identifier 'default.hudi_null_table', format 'hudi');
